+++ src/Components/ScopeUtils.js
